requirements.txt
python-telegram-bot>=13.0
twilio>=7.0.0
flask>=2.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
python-dotenv>=0.19.0 